requirements.txt
pyspark
sql-magic
nbqa
prefect 
azure-storage-file-datalake